improve import for split bill component
