--- components/SplitBill/SplitBill.tsx
@@ -1,5 +1,0 @@
-const SplitBill = () => {
-  console.log("Split Bill");
-};
-
-export default SplitBill;
